Add `rustc_shim_for_cargo_fix` and `wrapped_clippy_driver` to reuse code
